\name{FRGEpistasis-package}
\alias{FRGEpistasis-package}
\alias{FRGEpistasis}
\docType{package}
\title{
Package of Epistasis Detection by Functional Regression Model
}
\description{
FRGEpistasis is designed to detect the epistasis between genes or genomic regions for both common variants and rare variants. 
Currently FRGEpistasis was developed by Futao Zhang with R language and maintained in Xiong lab at UTSPH.
This tool is friendly, convenient and memory efficient. 
}
\details{
\tabular{ll}{
Package: \tab FRGEpistasis\cr
Type: \tab Package\cr
Version: \tab 0.99.5\cr
Date: \tab 2014-03-22\cr
License: \tab GPL-2\cr
}
It currently has the following functional modules: 
 Functional Regression Model (FRG) for Testing Interaction;
 Regression on Principal Components Analysis (PCA) for Testing Interaction; 
 Point-wise interaction Test.
}
\author{
Futao Zhang
Maintainer: Futao Zhang <futoaz@gmail.com>
}

\keyword{ package }
